requirements.txt
anyio==4.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.0.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
dash==3.0.4
dash-bootstrap-components==2.0.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
executing==0.8.3
fastjsonschema==2.20.0
Flask==3.0.3
flask-cors==5.0.1
Flask-Login==0.6.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httpie==3.2.4
httpx==0.27.0
idna==3.7
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.2
multidict==6.4.3
narwhals==1.38.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.5
overrides==7.4.0
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pip==25.1
platformdirs==4.3.7
plotly==6.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.43
psutil==5.9.0
psycopg2==2.9.10
pure-eval==0.2.2
pycparser==2.21
Pygments==2.19.1
pymongo==4.12.1
PyMySQL==1.1.1
PyQt6==6.7.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.1
referencing==0.30.2
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
Send2Trash==1.8.2
setuptools==80.1.0
sip==6.10.0
six==1.17.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
terminado==0.17.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
win_inet_pton==1.1.0
zipp==3.21.0
